requirements.txt
exa_py>=1.4.0
python-dotenv>=1.0.0
supabase>=2.12.0
